+++ examples/data-sources/nxos_port_channel_interface_member/data-source.tf
@@ -1,0 +1,4 @@
+data "nxos_port_channel_interface_member" "example" {
+  interface_id = "po1"
+  interface_dn = "sys/intf/phys-[eth1/11]"
+}