+++ examples/resources/nxos_port_channel_interface_member/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_port_channel_interface_member.example "sys/intf/aggr-[po1]/rsmbrIfs-[sys/intf/phys-[eth1/11]]"